@@ -1,78 +1,115 @@
 {
   "name": "ganadero",
-  "version": "0.0.0",
-  "author": "Serenapp",
-  "homepage": "http://ionicframework.com/",
-  "private": true,
+  "version": "0.0.1",
+  "author": "Ionic Framework",
+  "homepage": "https://ionicframework.com/",
   "scripts": {
-    "start": "ionic-app-scripts serve",
-    "clean": "ionic-app-scripts clean",
-    "build": "ionic-app-scripts build",
-    "lint": "ionic-app-scripts lint"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
+  "private": true,
   "dependencies": {
-	"auth0-js": "^9.7.3",
-	"@auth0/cordova": "^0.3.0",
-    "angular2-jwt": "^0.2.2",
-    "@angular/animations": "5.2.11",
-    "@angular/common": "5.2.11",
-    "@angular/compiler": "5.2.11",
-    "@angular/compiler-cli": "5.2.11",
-    "@angular/core": "5.2.11",
-    "@angular/forms": "5.2.11",
-    "@angular/http": "5.2.11",
-    "@angular/platform-browser": "5.2.11",
-    "@angular/platform-browser-dynamic": "5.2.11",
-    "@ionic-native/camera": "^4.11.0",
-    "@ionic-native/core": "~4.10.0",
-    "@ionic-native/diagnostic": "^4.11.0",
-    "@ionic-native/file-chooser": "^4.11.0",
-    "@ionic-native/file-path": "^4.11.0",
-    "@ionic-native/file-transfer": "^4.11.0",
-    "@ionic-native/splash-screen": "~4.10.0",
-    "@ionic-native/status-bar": "~4.10.0",
-    "@ionic/pro": "1.0.20",
-    "@ionic/storage": "2.1.3",
-    "angularfire2": "^5.0.0-rc.11",
-    "firebase": "^5.3.0",
-    "ionic-angular": "3.9.2",
-    "ionicons": "3.0.0",
-    "promise-polyfill": "^8.0.0",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
-    "sw-toolbox": "3.6.0",
-    "tesseract.js": "^1.0.10",
-    "zone.js": "0.8.26"
+    "@angular/common": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/fire": "^6.0.0",
+    "@angular/forms": "~8.2.14",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "@auth0/angular-jwt": "^4.0.0",
+    "@auth0/cordova": "^0.4.3",
+    "@capacitor/core": "2.0.0",
+    "@ionic-native/camera": "^5.23.0",
+    "@ionic-native/camera-preview": "^5.23.0",
+    "@ionic-native/core": "^5.0.7",
+    "@ionic-native/diagnostic": "^5.23.0",
+    "@ionic-native/file": "^5.23.0",
+    "@ionic-native/file-chooser": "^5.23.0",
+    "@ionic-native/file-path": "^5.23.0",
+    "@ionic-native/file-transfer": "^5.23.0",
+    "@ionic-native/in-app-browser": "^5.23.0",
+    "@ionic-native/safari-view-controller": "^5.23.0",
+    "@ionic-native/splash-screen": "^5.0.0",
+    "@ionic-native/status-bar": "^5.0.0",
+    "@ionic/angular": "^5.1.0-dev.202002181528.45d03ba",
+    "@ionic/pwa-elements": "^1.5.1",
+    "@ionic/storage": "^2.2.0",
+    "angularfire2": "^5.4.2",
+    "auth0-js": "^9.13.1",
+    "buffer": "^5.5.0",
+    "cordova-browser": "^6.0.0",
+    "cordova-plugin-camera": "^4.1.0",
+    "cordova-plugin-camera-preview": "^0.11.2",
+    "cordova-plugin-customurlscheme": "^5.0.1",
+    "cordova-plugin-file": "^6.0.2",
+    "cordova-plugin-file-transfer": "^1.7.1",
+    "cordova-plugin-filechooser": "^1.2.0",
+    "cordova-plugin-filepath": "^1.5.8",
+    "cordova-plugin-inappbrowser": "^3.2.0",
+    "cordova-plugin-safariviewcontroller": "^1.6.0",
+    "cordova.plugins.diagnostic": "^5.0.1",
+    "core-js": "^2.5.4",
+    "firebase": "^7.13.2",
+    "gulp-v4": "^4.0.0-alpha.5",
+    "rxjs": "~6.5.1",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "3.1.11",
-    "typescript": "~2.6.2"
+    "@angular-builders/custom-webpack": "^8.4.1",
+    "@angular-devkit/build-angular": "~0.803.24",
+    "@angular/cli": "~8.3.23",
+    "@angular/compiler": "~8.2.14",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@capacitor/cli": "2.0.0",
+    "@ionic/angular-toolkit": "^2.1.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^8.9.5",
+    "codelyzer": "^5.0.0",
+    "cordova-plugin-device": "^2.0.2",
+    "cordova-plugin-ionic-keyboard": "^2.2.0",
+    "cordova-plugin-ionic-webview": "^4.1.3",
+    "cordova-plugin-splashscreen": "^5.0.2",
+    "cordova-plugin-statusbar": "^2.4.2",
+    "cordova-plugin-whitelist": "^1.3.3",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   },
-  "cordovaPlugins": [
-    "cordova-plugin-device",
-    "cordova-plugin-console",
-    "cordova-plugin-splashscreen",
-    "cordova-plugin-whitelist",
-    "cordova-plugin-console",
-    "cordova-plugin-statusbar",
-    "cordova-plugin-device",
-    "cordova-plugin-splashscreen",
-    "ionic-plugin-keyboard"
-  ],
-  "cordovaPlatforms": [],
-  "description": "GanaderoNuevo: An Ionic project",
+  "description": "An Ionic project",
   "cordova": {
-    "platforms": [
-      "android",
-      "ios"
-    ],
     "plugins": {
-      "cordova-plugin-file-transfer": {},
-      "cordova-plugin-filechooser": {},
-      "cordova-plugin-filepath": {},
-      "cordova-plugin-camera": {},
-      "cordova.plugins.diagnostic": {},
-      "cordova-plugin-file": {}
-    }
+      "cordova-plugin-inappbrowser": {},
+      "cordova-plugin-whitelist": {},
+      "cordova-plugin-statusbar": {},
+      "cordova-plugin-device": {},
+      "cordova-plugin-splashscreen": {},
+      "cordova-plugin-ionic-webview": {},
+      "cordova-plugin-ionic-keyboard": {},
+      "cordova-plugin-safariviewcontroller": {},
+      "cordova-plugin-customurlscheme": {
+        "URL_SCHEME": "com.serenapp.ganadero",
+        "ANDROID_SCHEME": "com.serenapp.ganadero",
+        "ANDROID_HOST": "YOUR_DOMAIN",
+        "ANDROID_PATHPREFIX": "/cordova/com.serenapp.ganadero/callback"
+      }
+    },
+    "platforms": [
+      "browser"
+    ]
   }
 }
