@@ -18,48 +18,20 @@
  * specific language governing permissions and limitations
  * under the License.
  */
- 
-var path    = require('path'),
-    fs      = require('fs'),
-    shjs    = require('shelljs'),
-    zip     = require('adm-zip'),
-    Q       = require('q'),
-    clean   = require('./clean'),
-    check_reqs = require('./check_reqs'),
-    platformWwwDir          = path.join('platforms', 'browser', 'www'),
-    platformBuildDir        = path.join('platforms', 'browser', 'build'),
-    packageFile             = path.join(platformBuildDir, 'package.zip');
+
+var path = require('path');
+var check_reqs = require('./check_reqs');
 
 /**
  * run
  *   Creates a zip file int platform/build folder
  */
-module.exports.run = function(){
-
-    return check_reqs.run()
-    .then(function(){
-            return clean.cleanProject();
-        },
-        function checkReqsError(err){
-            console.error('Please make sure you meet the software requirements in order to build a browser cordova project');
-    })
-    .then(function(){
-
-        if (!fs.existsSync(platformBuildDir)) {
-            fs.mkdirSync(platformBuildDir);
-        }
-
-        // add the project to a zipfile
-        var zipFile = zip();
-        zipFile.addLocalFolder(platformWwwDir, '.');
-        zipFile.writeZip(packageFile);
-
-        return Q.resolve();
-
-    });
+module.exports.run = function () {
+    return check_reqs.run();
 };
 
-module.exports.help = function() {
+module.exports.help = function () {
     console.log('Usage: cordova build browser');
-    console.log('Build will create the packaged app in \''+platformBuildDir+'\'.');
+    var wwwPath = path.resolve(path.join(__dirname, '../../www'));
+    console.log("Build will create the packaged app in '" + wwwPath + "'.");
 };
