@@ -19,10 +19,9 @@ specific language governing permissions and limitations
 under the License.
 */
 
-//add methods as we determine what are the requirements
+// add methods as we determine what are the requirements
 
-var Q = require('q');
-
-module.exports.run = function() {
-    return Q.resolve();
+module.exports.run = function () {
+    // caller expects a promise resolved with an array of conditions
+    return Promise.resolve([]);
 };
